Refactored to get rid of TcpClient.Dispose() call
In older versions of .NET (System.Net.Sockets) the TcpClient class didn't
have a method for Dispose().  In order for this script to work in older
versions of .NET (and PowerShell) I needed to get rid of this call to
TcpClient.Dispose().  Instead, I've wrapped the error handling in so that
no matter what, I test for a connected TcpClient, and if so then just
close the connection.  This shouldn't have any remaining latches when the
script finishes.